Configuring the project and the Maven pom.xml
