--- DipBet_1.14/config.yml
@@ -1,121 +1,0 @@
-#  --= SuperBet =--
-# -Plugin by PurpleFishh
-# This plugin is a bet plugin with roulette game
-# This game is simple, you bet on a color(red, black or green), if you bet on red or black and you will get the money you bet x2 and if you bet on green and you win you will get x14 but it is a smaller chance 
-
-# - Commands -
-# /bet - give the player the item that opens the roulette inventory (superbet.command.itemgive)
-# /bet inventory - open the roulette inventory (superbet.command.itemgive.inventory)
-# /bet <player name> - give the item that opens the roulette inventory to an player (superbet.command.other)
-# /bet remove - remove the bet item from your inventory (superbet.command.remove)
-# /bet reload - reload plugin config (superbet.command.reload)
-# /bet help - the help command of the plugin (superbet.command.help)
-
-# You can use '&' and the color code after for color
-
-# - Options - 
-
-# Put the time that takes for the roulette to start (in sec)
-time: 30
-# To make the roulette run individual, for every player, not to have a global roulette for everyone 
-separate_roulette: false
-# If you want when the player close the Bet Inventory to be removed the Bet Item from their inventory
-delete_after_bet: false 
-# The amounts that you can bet
-bet1: 100
-bet2: 1000
-bet3: 10000
-bet4: 100000
-bet5: 1000000
-bet6: 10000000
-bet7: 100000000
-# How much the sum you bet will be multiplied, for each color, if you win 
-multiply_black: 2
-multiply_red: 2
-multiply_green: 14
-
-# Sound effects
-sound_effects: true
-# Fireworks for winners
-fireworks: true
-
-# Please use the text name of the items not id number, if you get an error please firstly verify the items id
-# Example: LIME_STAINED_GLASS(vers>1.13), STAINED_GLASS:5(vers<1.13)
-# -Betting Item
-
-# This are the informations for the betting item that the players right-click to open the inventory
-# Put the name of the item
-betting-item-name: "&6Super Bet"
-# Put the lore for the item
-betting-item-lore:
- - "&eWith this item you enter in the betting menu"
-# Put the item id
-betting-item-id: "GOLD_INGOT"
-
-# -Exit Item
-# That is the item that will be in every inventory for going back or exit
-exit_id: "SLIME_BALL"
-exit_name: "&cBack"
-
-# -Main bet inventory
-# This is the main inventory that will open when you right-click the betting item
-main_inv_name: "&cRoulette"
-
-# Betting Items: 
-red_id: "RED_TERRACOTTA"
-red_name: "&cRed"
-green_id: "GREEN_TERRACOTTA"
-green_name: "&aGreen"
-black_id: "BLACK_TERRACOTTA" 
-black_name: "&8Black"
-
-# Betting buttons: the items that if you click you will get in the inventory for betting on this color
-# You can use %bet_amount% - for the money amount already placed on that color
-red_button_name: "&cRed"
-red_button_lore: 
- - "&7 - Bet: &f%bet_amount%$"
-green_button_name: "&aGreen" 
-green_button_lore: 
- - "&7 - Bet: &f%bet_amount%$"
-black_button_name: "&8Black"
-black_button_lore: 
- - "&7 - Bet: &f%bet_amount%$"
-
-# Set a block in the inventory that will show the timer status(in how much seconds the game will start, if the timer is not started), (%time% - the time remained)
-status_no_bet_id: "EMERALD_BLOCK"
-status_no_bet: "&eNo one is betting"
-status_timer_id: "GOLD_BLOCK"
-status_timer: "&aThe game starts in: &b%time%"
-status_running_id: "REDSTONE_BLOCK"
-status_running: "&cThe game is running, please wait"
-
-# Set a item in the inventory for the last colors. This item will open an inventory where you will have the history of the colors that were in the last rounds
-last_colors_id: "MAP"
-last_colors_name: "&eLast colors"
-last_colors_lore:
- - "&7 - The colors from previous rounds"
-
-# The items that are on the background
-main_bg_id: "WHITE_STAINED_GLASS_PANE"
-line_bg_id: "GRAY_STAINED_GLASS_PANE"
-
-# -Bet money
-# This is the inventory that will open when you what to bet on a color and now you need to select the amount of money
-bet_inv_name: "&aBetting"
-
-# Increase and decrease the amount of money
-increase_id: "LIME_STAINED_GLASS"
-increase_amount_color: "&a" 
-decrease_id: "RED_STAINED_GLASS"
-decrease_amount_color: "&c" 
-
-# Put the bet or cancel it
-put_id: "DIAMOND"
-put_name: "&aBet"
-cancel_id: "SLIME_BALL"
-cance_name: "&cCancel"
-
-
-
-
-
