+++ plugin.yml
@@ -1,0 +1,10 @@
+name: SuperBet
+main:  me.purplefishh.dipcraft.superbet.main.Main
+version: 1.4.4
+author: PurpleFishh
+api-version: 1.13
+depend: [Vault]
+commands: 
+   bet:
+      description: Get Bet Item.
+      usage: /<command>
